Removed Client module, merged Types and Server modules in a single module.
