--- Types/src/main/java/module-info.java
@@ -1,4 +1,0 @@
-module stockmanager.types {
-	requires java.sql;
-	exports ch.stockmanager.types;
-}
