+++ src/main/java/ch/stockmanager/server/Server.java
