+++ src/main/java/ch/stockmanager/server/controllers/ProductController.java
