+++ src/main/java/ch/stockmanager/server/controllers/ShelfController.java
