+++ src/main/java/ch/stockmanager/server/util/ContextHelper.java
