+++ src/main/java/ch/stockmanager/server/util/DBInfo.java
