+++ src/main/java/ch/stockmanager/types/Lot.java
