+++ src/main/java/ch/stockmanager/types/Product.java
