+++ src/main/java/ch/stockmanager/types/ProductDateQuantity.java
