+++ src/main/java/ch/stockmanager/types/ProductShelf.java
