+++ src/main/java/ch/stockmanager/types/ProductYearSegment.java
