+++ src/main/java/ch/stockmanager/types/Sale.java
