+++ src/main/java/ch/stockmanager/types/Shelf.java
