+++ src/main/java/ch/stockmanager/types/Supplier.java
