+++ src/main/java/ch/stockmanager/types/YearSegment.java
