+++ src/main/resources/queries/lot.sql
