+++ src/main/resources/queries/product_default.sql
