+++ src/main/resources/queries/product_extra.sql
