+++ src/main/resources/queries/product_shelf_default.sql
