+++ src/main/resources/queries/product_year_segment.sql
