+++ src/main/resources/queries/sale_default.sql
