+++ src/main/resources/queries/shelf_default.sql
