+++ src/main/resources/queries/shelf_extra.sql
