+++ src/main/resources/queries/supplier_default.sql
