+++ src/main/resources/queries/supplier_extra.sql
