+++ src/main/resources/queries/year_segment.sql
